# sez/src/main/resources/api/swagger_full.json
{
  "openapi" : "3.0.3",
  "info" : {
    "description" : "API for managing the tutoring system (CRM)",
    "title" : "SEZ Tutoring Platform API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "description" : "Local Server",
    "url" : "http://localhost:8080"
  } ],
  "paths" : {
...
    "/api/v1/lessons" : {
      "post" : {
        "operationId" : "scheduleLesson",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateLessonDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Lesson scheduled successfully"
          }
        },
        "summary" : "Schedule a new lesson",
        "tags" : [ "Lessons" ]
      }
    },
    "/api/v1/lessons/my" : {
      "get" : {
        "operationId" : "getMyLessons",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : true,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : true,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "schema" : {
            "enum" : [ "PLANNED", "CONDUCTED", "NO_SHOW" ],
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/LessonResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "List of lessons retrieved"
          }
        },
        "summary" : "Get lessons for current user (calendar view)",
        "tags" : [ "Lessons" ]
      }
    },
    "/api/v1/lessons/{id}" : {
      "delete" : {
        "description" : "Removes a scheduled lesson from the system.",
        "operationId" : "deleteLesson",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Lesson deleted successfully"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden - only tutors can delete lessons"
          },
          "404" : {
            "description" : "Lesson not found"
          }
        },
        "summary" : "Delete a lesson by ID",
        "tags" : [ "Lessons" ]
      },
      "patch" : {
        "operationId" : "updateLesson",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateLessonDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Lesson updated successfully"
          },
          "404" : {
            "description" : "Lesson not found"
          }
        },
        "summary" : "Update lesson details (time, status, or meeting link)",
        "tags" : [ "Lessons" ]
      }
    },
    "/lessons/{id}/details" : {
      "get" : {
        "operationId" : "getLessonDetails",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LessonDetailsResponse"
                }
              }
            },
            "description" : "Full lesson details retrieved"
          },
          "404" : {
            "description" : "Lesson not found"
          }
        },
        "summary" : "Get full lesson details including tutor, student and course info",
        "tags" : [ "Lessons" ]
      }
    },
...
    "/api/v1/reviews" : {
      "get" : {
        "operationId" : "getTutorReviews",
        "parameters" : [ {
          "in" : "query",
          "name" : "tutorId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "sortByRatingUp",
          "schema" : {
            "default" : false,
            "description" : "Sort reviews by rating in descending order",
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "schema" : {
            "default" : 20,
            "maximum" : 20,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ReviewResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "List of reviews retrieved successfully"
          }
        },
        "summary" : "Get list of reviews for a specific tutor",
        "tags" : [ "Reviews" ]
      },
      "post" : {
        "operationId" : "createReview",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateReviewDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Review created successfully"
          }
        },
        "summary" : "Create a new review for a tutor",
        "tags" : [ "Reviews" ]
      }
    },
    "/api/v1/reviews/my" : {
      "get" : {
        "operationId" : "getMyReviewForTutor",
        "parameters" : [ {
          "in" : "query",
          "name" : "tutorId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description" : "Review retrieved successfully"
          },
          "404" : {
            "description" : "Review not found"
          }
        },
        "summary" : "Get current student's review for a specific tutor",
        "tags" : [ "Reviews" ]
      }
    },
    "/api/v1/reviews/{id}" : {
      "delete" : {
        "operationId" : "deleteReview",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Review deleted successfully"
          },
          "404" : {
            "description" : "Review not found"
          }
        },
        "summary" : "Delete review by ID",
        "tags" : [ "Reviews" ]
      },
      "patch" : {
        "operationId" : "updateReview",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateReviewDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Review updated successfully"
          },
          "404" : {
            "description" : "Review not found"
          }
        },
        "summary" : "Update existing review rating and comment",
        "tags" : [ "Reviews" ]
      }
    },
    "/api/v1/notifications/announcements" : {
      "get" : {
        "operationId" : "getAnnouncements",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "schema" : {
            "default" : 20,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AnnouncementResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "List of announcements"
          }
        },
        "summary" : "Get only informational notifications (announcements)",
        "tags" : [ "Notifications" ]
      }
    },
    "/api/v1/notifications/requests" : {
      "get" : {
        "operationId" : "getActionRequests",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "schema" : {
            "default" : 20,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ActionRequestResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "List of action requests"
          }
        },
        "summary" : "Get only action requests",
        "tags" : [ "Notifications" ]
      },
      "post" : {
        "operationId" : "createActionRequest",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateActionRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          }
        },
        "summary" : "Create a new action request",
        "tags" : [ "Notifications" ]
      }
    },
    "/api/v1/notifications/{id}" : {
      "delete" : {
        "operationId" : "deleteNotification",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          }
        },
        "summary" : "Delete notification",
        "tags" : [ "Notifications" ]
      },
      "patch" : {
        "operationId" : "markAsRead",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Marked as read"
          }
        },
        "summary" : "Mark notification as read",
        "tags" : [ "Notifications" ]
      }
    },
    "/api/v1/notifications/{id}/handle" : {
      "post" : {
        "operationId" : "handleNotificationAction",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NotificationHandleDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "summary" : "Accept or Decline an action request",
        "tags" : [ "Notifications" ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "StudentResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "email" : {
            "format" : "email",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "role" : {
            "enum" : [ "STUDENT" ],
            "type" : "string"
          }
        },
        "required" : [ "email", "id", "name", "role", "surname" ],
        "type" : "object"
      },
...
      "CourseResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "tutorName" : {
            "type" : "string"
          },
          "tutorSurname" : {
            "type" : "string"
          },
          "tutorId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "subjectName" : {
            "type" : "string"
          },
          "hourlyRate" : {
            "type" : "number"
          },
          "description" : {
            "type" : "string"
          },
          "rating" : {
            "maximum" : 5,
            "minimum" : 1,
            "type" : "number"
          }
        },
        "required" : [ "description", "hourlyRate", "id", "rating", "subjectName", "tutorId", "tutorName", "tutorSurname" ],
        "type" : "object"
      },
...
      "EnrollmentDetailsResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "REQUESTED", "ACTIVE" ],
            "type" : "string"
          },
          "student" : {
            "$ref" : "#/components/schemas/StudentResponse"
          },
          "course" : {
            "$ref" : "#/components/schemas/CourseResponse"
          }
        },
        "required" : [ "course", "id", "status", "student" ],
        "type" : "object"
      },
      "CreateLessonDto" : {
        "properties" : {
          "enrollmentId" : {
            "description" : "ID запису на курс (Enrollment)",
            "format" : "uuid",
            "type" : "string"
          },
          "startTime" : {
            "description" : "Час початку уроку. Повинен бути у майбутньому.",
            "format" : "date-time",
            "type" : "string"
          },
          "endTime" : {
            "description" : "Час закінчення уроку. Повинен бути після часу початку.",
            "format" : "date-time",
            "type" : "string"
          },
          "price" : {
            "description" : "Ціна за урок. Не може бути від’ємною.",
            "minimum" : 0,
            "type" : "number"
          },
          "videoCallUrl" : {
            "description" : "Посилання на відеодзвінок.",
            "format" : "uri",
            "type" : "string"
          }
        },
        "required" : [ "endTime", "enrollmentId", "startTime" ],
        "type" : "object"
      },
      "LessonResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "startTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "endTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/LessonStatus"
          },
          "courseName" : {
            "type" : "string"
          },
          "subjectName" : {
            "type" : "string"
          },
          "partnerName" : {
            "type" : "string"
          },
          "videoCallUrl" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "required" : [ "endTime", "id", "startTime", "status" ],
        "type" : "object"
      },
      "LessonStatus" : {
        "enum" : [ "PLANNED", "CONDUCTED", "NO_SHOW" ],
        "type" : "string"
      },
      "UpdateLessonDto" : {
        "description" : "Поля для часткового оновлення уроку. При зміні часу проводиться перевірка на наявність конфліктів у розкладі.",
        "properties" : {
          "startTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "endTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "videoCallUrl" : {
            "format" : "uri",
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/LessonStatus"
          }
        },
        "type" : "object"
      },
      "LessonDetailsResponse" : {
        "properties" : {
          "lesson" : {
            "$ref" : "#/components/schemas/LessonResponse"
          },
          "enrollmentDetails" : {
            "$ref" : "#/components/schemas/EnrollmentDetailsResponse"
          }
        },
        "type" : "object"
      },
...
      "ReviewResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "tutorId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "studentId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "tutorName" : {
            "description" : "Name of the tutor being reviewed",
            "type" : "string"
          },
          "tutorSurname" : {
            "description" : "Surname of the tutor being reviewed",
            "type" : "string"
          },
          "studentName" : {
            "description" : "Name of the student who wrote the review",
            "type" : "string"
          },
          "studentSurname" : {
            "description" : "Surname of the student who wrote the review",
            "type" : "string"
          },
          "rating" : {
            "maximum" : 5,
            "minimum" : 1,
            "type" : "integer"
          },
          "comment" : {
            "description" : "Detailed feedback from student",
            "type" : "string"
          }
        },
        "required" : [ "comment", "id", "rating", "studentId", "tutorId" ],
        "type" : "object"
      },
      "CreateReviewDto" : {
        "properties" : {
          "tutorId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "rating" : {
            "description" : "Rating from 1 to 5 stars",
            "maximum" : 5,
            "minimum" : 1,
            "type" : "integer"
          },
          "comment" : {
            "description" : "Detailed feedback about the course",
            "maxLength" : 500,
            "minLength" : 10,
            "type" : "string"
          }
        },
        "required" : [ "rating", "tutorId" ],
        "type" : "object"
      },
      "UpdateReviewDto" : {
        "properties" : {
          "rating" : {
            "description" : "Rating from 1 to 5 stars",
            "maximum" : 5,
            "minimum" : 1,
            "type" : "integer"
          },
          "comment" : {
            "description" : "Detailed feedback about the course",
            "maxLength" : 500,
            "minLength" : 10,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AnnouncementResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "isRead" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "createdAt", "id", "isRead", "message" ],
        "type" : "object"
      },
      "ActionRequestResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "isRead" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "actionType" : {
            "$ref" : "#/components/schemas/NotificationActionType"
          },
          "relatedEntityId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "completed" : {
            "type" : "boolean"
          },
          "metadata" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "type" : "object"
          }
        },
        "required" : [ "actionType", "completed", "createdAt", "id", "isRead", "message", "relatedEntityId" ],
        "type" : "object"
      },
      "NotificationActionType" : {
        "enum" : [ "ENROLLMENT_CONFIRM", "LESSON_RESCHEDULE", "LESSON_CANCEL" ],
        "type" : "string"
      },
      "CreateActionRequestDto" : {
        "properties" : {
          "recipientId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "actionType" : {
            "$ref" : "#/components/schemas/NotificationActionType"
          },
          "relatedEntityId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "metadata" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "type" : "object"
          }
        },
        "required" : [ "actionType", "message", "recipientId", "relatedEntityId" ],
        "type" : "object"
      },
      "NotificationHandleDto" : {
        "properties" : {
          "confirmed" : {
            "type" : "boolean"
          }
        },
        "required" : [ "confirmed" ],
        "type" : "object"
      },
...
    }
  }
}